@@ -10,13 +10,13 @@ authors = [
 ]
 description = "Spatial image analysis with pytorch and caffe backends."
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 dependencies = [
     "Pillow>=10.2.0",
     "numpy>=1.19",
     "scikit-learn==1.1.3",
-    "torch==1.13.1",
-    "torchvision==0.14.1",
+    "torch>=1.13.1,<2.3",
+    "torchvision>=0.14.1,<0.18",
     "boto3",
 ]
 license = {file = "license.txt"}
