@@ -343,9 +343,7 @@ class ClassifyReturnMsgTest(unittest.TestCase):
                 self.assertIsNone(col)
 
         for class_ in return_msg.classes:
-            # Change to LabelID when on Python 3.10+ only.
-            self.assertTrue(
-                isinstance(class_, int) or isinstance(class_, str))
+            self.assertTrue(isinstance(class_, LabelId))
 
         self.assertTrue(isinstance(return_msg.valid_rowcol, bool))
 
